+++ src/lib.rs
@@ -83,6 +83,7 @@
 use env::{EnvNoProxy, EnvProxies};
 use url::Url;
 
+mod noproxy;
 mod types;
 
 pub mod env;
@@ -90,10 +91,14 @@ pub mod env;
 #[cfg(all(unix, not(target_os = "mac_os")))]
 pub mod unix;
 
+pub use noproxy::NoProxyResolver;
 pub use types::ProxyResolver;
 
 #[cfg(all(unix, not(target_os = "mac_os")))]
 use unix::UnixProxyResolver as SystemProxyResolverImpl;
+
+#[cfg(windows)]
+use NoProxyResolver as SystemProxyResolverImpl;
 
 /// The system proxy resolver.
 ///